+++ core/tests/transition_matrix/mod.rs
@@ -10,4 +10,3 @@ pub mod helpers;
 pub mod initial_invalid_tests;
 pub mod pinging_invalid_tests;
 pub mod state_transition_matrix_tests;
-pub mod timed_out_invalid_tests;